move no_debugging_tables out of googlefonts directory
