+++ tests/test_checks_googlefonts.py
@@ -3745,17 +3745,6 @@ def test_check_meta_script_lang_tags():
     assert_results_contain(check(ttFont), WARN, "lacks-meta-table")
 
 
-def test_check_no_debugging_tables():
-    """Ensure fonts do not contain any preproduction tables."""
-    check = CheckTester("no_debugging_tables")
-
-    ttFont = TTFont(TEST_FILE("overpassmono/OverpassMono-Regular.ttf"))
-    assert_results_contain(check(ttFont), WARN, "has-debugging-tables")
-
-    del ttFont["FFTM"]
-    assert_PASS(check(ttFont))
-
-
 def test_check_metadata_family_directory_name():
     """Check family directory name."""
     check = CheckTester("googlefonts:metadata/family_directory_name")
